+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 1.1
+- Add Disable REST API user endpoints
+- Add link to "Disable REST API"
+- Add a submit button on top of setting admin page
+
 ## 1.01
 - Add wptuning.png for [Entrepot](https://github.com/imath/entrepot/ "Entrepot") by imath